Fix warnings produced during building

+++ src/ale/games/supported/MrDo.cpp
@@ -102,7 +102,7 @@ void MrDoSettings::setMode(
     std::unique_ptr<StellaEnvironmentWrapper> environment) {
   if (m < 4) {
     // Press select until the correct mode is reached.
-    while (readRam(&system, 0x80) != m) { environment->pressSelect(5); }
+    while (readRam(&system, 0x80) != static_cast<int>(m)) { environment->pressSelect(5); }
 
     // Reset the environment to apply changes.
     environment->softReset();
